Extract SimpleLoggingSoapHandler from SOAP handlers
* Extract common SOAPHandler code into SimpleLoggingSoapHandler
* Refactor WsdlFirstServiceHandler and WsdlFirstClientHandler
  to extend SimpleLoggingSoapHandler
