+++ dropwizard-jaxws-example/src/main/java/com/roskart/dropwizard/jaxws/example/util/SimpleLoggingSoapHandler.java
@@ -1,0 +1,50 @@
+package com.roskart.dropwizard.jaxws.example.util;
+
+import jakarta.xml.ws.handler.MessageContext;
+import jakarta.xml.ws.handler.soap.SOAPHandler;
+import jakarta.xml.ws.handler.soap.SOAPMessageContext;
+import org.slf4j.Logger;
+
+import java.util.Set;
+import javax.xml.namespace.QName;
+
+public abstract class SimpleLoggingSoapHandler implements SOAPHandler<SOAPMessageContext> {
+
+    private final String name;
+    private final Logger logger;
+
+    protected SimpleLoggingSoapHandler(String name, Logger logger) {
+        this.name = name;
+        this.logger = logger;
+    }
+
+    @Override
+    public Set<QName> getHeaders() {
+        return Set.of();
+    }
+
+    @Override
+    public void close(MessageContext messageContext) {
+        logger.info("Closing");
+    }
+
+    @Override
+    public boolean handleFault(SOAPMessageContext context) {
+        logger.warn("Handling fault");
+        return true;
+    }
+
+    @Override
+    public boolean handleMessage(SOAPMessageContext context) {
+        var outbound = (Boolean) context.get(MessageContext.MESSAGE_OUTBOUND_PROPERTY);
+
+        if (Boolean.TRUE.equals(outbound)) {
+            logger.info("{} - outbound", name);
+        } else {
+            logger.info("{} - inbound", name);
+        }
+
+        return true;
+    }
+
+}
