+++ dropwizard-jaxws-example/src/main/java/com/roskart/dropwizard/jaxws/example/ws/WsdlFirstServiceHandler.java
@@ -1,46 +1,19 @@
 package com.roskart.dropwizard.jaxws.example.ws;
 
-import jakarta.xml.ws.handler.MessageContext;
-import jakarta.xml.ws.handler.soap.SOAPHandler;
-import jakarta.xml.ws.handler.soap.SOAPMessageContext;
+import com.roskart.dropwizard.jaxws.example.util.SimpleLoggingSoapHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.namespace.QName;
-import java.util.Set;
-
-public class WsdlFirstServiceHandler implements SOAPHandler<SOAPMessageContext> {
+/**
+ * Jakarta XML Web Services server handler when WsdlFirstService is invoked.
+ *
+ * See 
+ */
+public class WsdlFirstServiceHandler extends SimpleLoggingSoapHandler {
 
     private static final Logger LOG = LoggerFactory.getLogger(WsdlFirstServiceHandler.class);
 
-    @Override
-    public Set<QName> getHeaders() {
-        return Set.of();
+    public WsdlFirstServiceHandler() {
+        super("WsdlFirstService server handler", LOG);
     }
-
-    @Override
-    public void close(MessageContext messageContext) {
-        LOG.info("Closing");
-    }
-
-    @Override
-    public boolean handleFault(SOAPMessageContext context) {
-        LOG.warn("Handling fault");
-        return true;
-    }
-
-    @Override
-    public boolean handleMessage(SOAPMessageContext context) {
-
-        Boolean outbound = (Boolean) context.get(MessageContext.MESSAGE_OUTBOUND_PROPERTY);
-
-        if (Boolean.TRUE.equals(outbound)) {
-            LOG.info("WsdlFirstService server handler - outbound");
-        } else {
-            LOG.info("WsdlFirstService server handler - inbound");
-        }
-
-        return true;
-    }
-
 }
